Convert Rageshake to an interface for testing purpose.
